fix: Include credentials in robotoff read requests (#265)
Because the robotoff response can depend on this.

For example
https://robotoff.openfoodfacts.org/api/v1/questions/3336971407225?lang=en
(at the time of writing) return an insight only if you are
authenticated.

(it might be, in this case, because we have a skip by DEVICE_ID as the
question was voted already, but if I'm logged in, it should be skip by
USER).